fix: converge worker session history replay validation
