+++ docs/internal/baselines/backend-exemption-budget.json
@@ -1755,12 +1755,6 @@
     },
     {
       "rule": "pkgmaintcheck:ignore-cyclomatic-complexity",
-      "target": "pkg/services/recordings/internal/projections/projectiontests/world_state_payload_support_test.go#assignGeneratedProjectionPayload",
-      "owner": "backend-maintainers",
-      "removalReason": "refactor this target and remove this exemption when the threshold debt is resolved."
-    },
-    {
-      "rule": "pkgmaintcheck:ignore-cyclomatic-complexity",
       "target": "pkg/services/recordings/internal/projections/projectiontests/world_state_support_test.go#assignGeneratedProjectionSessionLifecyclePayload",
       "owner": "backend-maintainers",
       "removalReason": "refactor this target and remove this exemption when the threshold debt is resolved."
@@ -1768,6 +1762,12 @@
     {
       "rule": "pkgmaintcheck:ignore-cyclomatic-complexity",
       "target": "pkg/services/recordings/internal/projections/projectiontests/world_state_test.go#TestReconstructFactoryWorldState_PreservesSafeInferenceAttemptDiagnostics",
+      "owner": "backend-maintainers",
+      "removalReason": "refactor this target and remove this exemption when the threshold debt is resolved."
+    },
+    {
+      "rule": "pkgmaintcheck:ignore-cyclomatic-complexity",
+      "target": "pkg/services/recordings/internal/projections/projectiontests/world_view_support_test.go#assignGeneratedProjectionPayload",
       "owner": "backend-maintainers",
       "removalReason": "refactor this target and remove this exemption when the threshold debt is resolved."
     },
